fix bugs

+++ utils/http.py
@@ -29,7 +29,7 @@ def encode_datetime(obj):
 class DateTimeEncoder(json.JSONEncoder):
     def default(self, o):
         if isinstance(o, datetime):
-            return (o - epoch).total_seconds() * 1000
+            return int((o - epoch).total_seconds() * 1000)
         elif isinstance(o, date):
             return o.strftime('%Y-%m-%d')
         elif isinstance(o, uuid.UUID):
